+++ fake-cucumber/javascript/src/cli.ts
@@ -2,7 +2,8 @@ import { Command } from 'commander'
 import packageJson from '../package.json'
 import loadSupportCode from './loadSupportCode'
 import runCucumber from './runCucumber'
-import gherkin, {
+import {
+  GherkinStreams,
   IGherkinOptions,
   Query as GherkinQuery,
 } from '@cucumber/gherkin'
@@ -43,7 +44,7 @@ async function main() {
     createReadStream: (path: string) =>
       fs.createReadStream(path, { encoding: 'utf-8' }),
   }
-  const gherkinEnvelopeStream = gherkin.fromPaths(paths, gherkinOptions)
+  const gherkinEnvelopeStream = GherkinStreams.fromPaths(paths, gherkinOptions)
 
   const envelopeOutputStream = makeFormatStream(format)
   envelopeOutputStream.pipe(process.stdout)