+++ gherkin/javascript/src/Dialect.ts
@@ -1,15 +1,15 @@
 export default interface Dialect {
-  name: string
-  native: string
-  feature: readonly string[]
-  background: readonly string[]
-  rule: readonly string[]
-  scenario: readonly string[]
-  scenarioOutline: readonly string[]
-  examples: readonly string[]
-  given: readonly string[]
-  when: readonly string[]
-  then: readonly string[]
-  and: readonly string[]
-  but: readonly string[]
+  readonly name: string
+  readonly native: string
+  readonly feature: ReadonlyArray<string>
+  readonly background: ReadonlyArray<string>
+  readonly rule: ReadonlyArray<string>
+  readonly scenario: ReadonlyArray<string>
+  readonly scenarioOutline: ReadonlyArray<string>
+  readonly examples: ReadonlyArray<string>
+  readonly given: ReadonlyArray<string>
+  readonly when: ReadonlyArray<string>
+  readonly then: ReadonlyArray<string>
+  readonly and: ReadonlyArray<string>
+  readonly but: ReadonlyArray<string>
 }